First Commit.
Code generates a working prompt of different colors (red/blue) to chat with the Master Control Program via your terminal. It uses OpenAI to generate a response to questions from the user. 'quit' is used to end the chat. Currently the color scheme is messed up when compiled for windows so that needs to be fixed.